+++ apps/dashboard/src/app/(main)/(protected)/projects/[projectId]/users/page-client.tsx
@@ -11,18 +11,18 @@ import { useAdminApp } from "../use-admin-app";
 export default function PageClient() {
   const stackAdminApp = useAdminApp();
   const project = stackAdminApp.useProject();
-  const users = stackAdminApp.useUsers();
+  const firstUser = stackAdminApp.useUsers({ limit: 1 });
 
   return (
     <PageLayout
       title="Users"
-      description={`User count: ${project.userCount}`}
+      description={`Total: ${project.userCount}`}
       actions={<UserDialog
         type="create"
         trigger={<Button>Create User</Button>}
       />}
     >
-      {users.length > 0 ? null : (
+      {firstUser.length > 0 ? null : (
         <Alert variant='success'>
           Congratulations on starting your project! Check the <StyledLink href="https://docs.stack-auth.com">documentation</StyledLink> to add your first users.
         </Alert>