+++ apps/oauth-mock-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/oauth-mock-server
 
+## 2.6.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.30
+
 ## 2.6.29
 
 ### Patch Changes