+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/example-demo-app
 
+## 2.6.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.30
+  - @stackframe/stack@2.6.30
+  - @stackframe/stack-ui@2.6.30
+
 ## 2.6.29
 
 ### Patch Changes